Decided to move BasicStop to its own class (and have the actual stop extend it)
